Replaced MediatR event bus with the custom one to enable scoping, tracing etc. Plugged that fully into Marten and EventStoreDB
