+++ Core.Marten/Subscriptions/MartenSubscription.cs
@@ -1,16 +1,22 @@
 ﻿using Marten;
 using Marten.Events;
 using Marten.Events.Projections;
+using Microsoft.Extensions.Logging;
 
 namespace Core.Marten.Subscriptions;
 
 public class MartenSubscription: IProjection
 {
     private readonly IEnumerable<IMartenEventsConsumer> consumers;
+    private readonly ILogger<MartenSubscription> logger;
 
-    public MartenSubscription(IEnumerable<IMartenEventsConsumer> consumers)
+    public MartenSubscription(
+        IEnumerable<IMartenEventsConsumer> consumers,
+        ILogger<MartenSubscription> logger
+    )
     {
         this.consumers = consumers;
+        this.logger = logger;
     }
 
     public void Apply(
@@ -25,13 +31,20 @@ public class MartenSubscription: IProjection
         CancellationToken ct
     )
     {
-        foreach (var consumer in consumers)
+        try
         {
-            await consumer.ConsumeAsync(operations, streams, ct);
+            foreach (var consumer in consumers)
+            {
+                await consumer.ConsumeAsync(operations, streams, ct);
+            }
+        }
+        catch (Exception exc)
+        {
+            logger.LogError("Error while processing Marten Subscription: {ExceptionMessage}", exc.Message);
+            throw;
         }
     }
 }
-
 
 public interface IMartenEventsConsumer
 {
