+++ Core.WebApi/Tracing/TracingMiddleware.cs
@@ -1,10 +1,7 @@
 ﻿using Core.Tracing;
-using Core.Tracing.Causation;
-using Core.Tracing.Correlation;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Logging;
 
 namespace Core.WebApi.Tracing;
