+++ Core/Events/IEventHandler.cs
@@ -1,8 +1,6 @@
-using MediatR;
-
 namespace Core.Events;
 
-public interface IEventHandler<in TEvent>: INotificationHandler<TEvent>
-    where TEvent : IEvent
+public interface IEventHandler<in TEvent>
 {
+    Task Handle(TEvent @event, CancellationToken ct);
 }
