--- Core/Events/NoMediator/INoMediatorEventHandler.cs
@@ -1,6 +1,0 @@
-namespace Core.Events.NoMediator;
-
-public interface INoMediatorEventHandler<in TEvent>
-{
-    Task Handle(TEvent @event, CancellationToken ct);
-}
