+++ Sample/EventPipelines/EventPipelines.Tests/ClassesWithBuilderTest.cs
