+++ Sample/EventPipelines/EventPipelines.Tests/ClassesWithIoCTest.cs
