+++ Sample/EventPipelines/EventPipelines.Tests/ClassesWithMediatRTest.cs
