+++ Sample/EventPipelines/EventPipelines/Configuration.cs
