+++ Sample/EventStoreDB/ECommerce/Carts/Carts/ShoppingCarts/Config.cs
@@ -8,7 +8,6 @@ using Carts.ShoppingCarts.GettingCarts;
 using Carts.ShoppingCarts.InitializingCart;
 using Carts.ShoppingCarts.RemovingProduct;
 using Core.Commands;
-using Core.Events.NoMediator;
 using Core.EventStoreDB.Repository;
 using Core.Marten.ExternalProjections;
 using Core.Queries;
@@ -35,14 +34,6 @@ internal static class CartsConfig
             .AddCommandHandler<AddProduct, HandleAddProduct>()
             .AddCommandHandler<RemoveProduct, HandleRemoveProduct>()
             .AddCommandHandler<ConfirmShoppingCart, HandleConfirmCart>();
-    }
-
-    public class MyClass: INoMediatorEventHandler<ShoppingCartInitialized>
-    {
-        public Task Handle(ShoppingCartInitialized @event, CancellationToken ct)
-        {
-            return Task.CompletedTask;
-        }
     }
 
     private static IServiceCollection AddProjections(this IServiceCollection services)
