+++ Sample/MeetingsManagement/MeetingsManagement/Notifications/Config.cs
@@ -1,4 +1,4 @@
-using MediatR;
+using Core.Events;
 using MeetingsManagement.Meetings.CreatingMeeting;
 using MeetingsManagement.Notifications.NotifyingByEvent;
 using Microsoft.Extensions.DependencyInjection;
@@ -8,5 +8,5 @@ namespace MeetingsManagement.Notifications;
 public static class Config
 {
     public static IServiceCollection AddNotifications(this IServiceCollection services) =>
-        services.AddScoped<INotificationHandler<MeetingCreated>, EmailNotifier>();
+        services.AddEventHandler<MeetingCreated, EmailNotifier>();
 }
