+++ Sample/MeetingsManagement/MeetingsManagement/Notifications/NotifyingByEvent/EmailNotifier.cs
